+++ CHANGELOG.md
@@ -13,6 +13,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added an initial `benchmark` suite to be run by `npm run benchmarks`.
+
 ## [11.1.2] - 2018-09-19
 
 ### Changed